feat: generalize dual Python IPC and speed-up conda builds
